requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
discord.py==2.5.2
distro==1.9.0
et_xmlfile==2.0.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.7.0
greenlet==3.2.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
multidict==6.5.0
numpy==2.2.6
openai==1.88.0
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pefile==2023.2.7
pillow==11.1.0
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-steam-api==2.0.4
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.3
setuptools==80.7.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.41
tqdm==4.67.1
ttkthemes==3.2.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
XlsxWriter==3.2.3
yarl==1.20.1
